mlbackend/requirements-minimal.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
joblib
deep-translator
requests
httpx
pydantic-settings
